IS-0: Move common items to common_schedule module (to improve build speed)
