+++ modules/common_schedule/src/main/java/kekmech/ru/common_schedule/items/SelfStudyItem.kt
@@ -1,25 +1,24 @@
-package kekmech.ru.feature_schedule.main.item
+package kekmech.ru.common_schedule.items
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import kekmech.ru.common_adapter.AdapterItem
 import kekmech.ru.common_adapter.BaseItemBinder
-import kekmech.ru.feature_schedule.R
+import kekmech.ru.common_schedule.R
 
-internal object SelfStudyItem
+object SelfStudyItem
 
-internal interface SelfStudyViewHolder
+interface SelfStudyViewHolder
 
-internal class SelfStudyViewHolderImpl(
+private class SelfStudyViewHolderImpl(
     containerView: View
 ) : SelfStudyViewHolder, RecyclerView.ViewHolder(containerView)
 
-internal class SelfStudyItemBinder : BaseItemBinder<SelfStudyViewHolder, SelfStudyItem>() {
-
+private class SelfStudyItemBinder : BaseItemBinder<SelfStudyViewHolder, SelfStudyItem>() {
     override fun bind(vh: SelfStudyViewHolder, model: SelfStudyItem, position: Int) = Unit
 }
 
-internal class SelfStudyAdapterItem : AdapterItem<SelfStudyViewHolder, SelfStudyItem>(
+class SelfStudyAdapterItem : AdapterItem<SelfStudyViewHolder, SelfStudyItem>(
     isType = { it is SelfStudyItem },
     layoutRes = R.layout.item_self_study,
     viewHolderGenerator = ::SelfStudyViewHolderImpl,
