+++ modules/common_schedule/src/main/res/drawable/ic_self_study_day.xml
