+++ modules/common_schedule/src/main/res/layout/item_self_study.xml
