+++ modules/feature_schedule/src/main/java/kekmech/ru/feature_schedule/main/item/WorkingDayItem.kt
@@ -8,6 +8,8 @@ import kekmech.ru.common_adapter.BaseAdapter
 import kekmech.ru.common_adapter.BaseItemBinder
 import kekmech.ru.common_android.openLinkExternal
 import kekmech.ru.common_schedule.items.ClassesAdapterItem
+import kekmech.ru.common_schedule.items.LunchAdapterItem
+import kekmech.ru.common_schedule.items.SelfStudyAdapterItem
 import kekmech.ru.coreui.items.EmptyStateAdapterItem
 import kekmech.ru.coreui.items.NotePreviewAdapterItem
 import kekmech.ru.coreui.items.SpaceAdapterItem
